Set up Batch files for testing. Removed most of old comments

+++ tools/Benchmarks/test.bat
@@ -1,11 +1,28 @@
-rem test.bat "solver.exe" level_from level_to
-
 @echo off
+
+if [%1] == [] goto printUsage
+if [%2] == [] goto printUsage
+if [%3] == [] goto printUsage
+
+
 for /l %%i in (%2, 1, %3) do (
-rem for /l %%i in (0, 1, 100) do (
 	< nul set /p a="Level %%i:	"
-	copy "levels\Level%%i" "mortal_coil.txt" > nul
+	copy "..\..\data\levels\Level%%i" "mortal_coil.txt" > nul
 	if exist "output.txt" del /q "output.txt"
-	"%1" > "times\time%%i"
-	if exist "output.txt" "..\Solution Tester.exe" < "output.txt"
+
+	for %%e in (%1) do (
+		start "" /d "%cd%" /b /w "%%e" > "times\time%%i"
+	)
+	
+	if exist "output.txt" "..\Solution Tester\Solution Tester.exe" < "output.txt"
 )
+
+rem if exist "output.txt" del /q "output.txt"
+rem if exist "mortal_coil.txt" del /q "mortal_coil.txt"
+goto end
+
+:printUsage
+echo Usage: test.bat "solver.exe" levelFrom levelTo
+goto end
+
+:end
